feat(user-signup): Refactor to fix code duplication
