+++ app/main/auth/user.py
@@ -1,30 +1,5 @@
 from flask import jsonify
-from app.models import User, Owner
 from app.db import users
-
-
-def create_store_attendant(username, email, password):
-    """
-    This method creates a new store attendant.
-    parameters: username, email, password.
-    returns: a success message.
-    """
-    attendant = User(username=username, email=email, password=password)
-    users.append(attendant)
-    return jsonify({
-        "message": "Store attendant account successfully created."
-    }), 201
-
-
-def create_store_owner(username, email, password):
-    """
-    This method creates a new store owner or admin.
-    parameters: username, email, password.
-    returns: a success message.
-    """
-    owner = Owner(username=username, email=email, password=password)
-    users.append(owner)
-    return jsonify({"message": "Account successfully created."}), 201
 
 
 def get_all_users():
